refactor rcnn metrics
